+++ src/irvine/oeis/a068/A068607.java
@@ -17,6 +17,6 @@ public class A068607 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return Z.valueOf(n * k).multiply(n + k + 1);
+    return Z.valueOf(n * (long) k).multiply(n + k + 1);
   }
 }
