+++ src/irvine/oeis/a100/A100345.java
@@ -17,6 +17,6 @@ public class A100345 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return Z.valueOf(n * (n + k));
+    return Z.valueOf(n * (long) (n + k));
   }
 }
