+++ src/irvine/oeis/a132/A132111.java
@@ -17,6 +17,6 @@ public class A132111 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return Z.valueOf(n).square().add(Z.valueOf(k * (n + k)));
+    return Z.valueOf(n).square().add(Z.valueOf(k * (long) (n + k)));
   }
 }
