+++ src/irvine/oeis/a132/A132121.java
@@ -17,6 +17,6 @@ public class A132121 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return Z.valueOf(n * (n + 1)).multiply((3 * k + 2) * n + 1).divide(6);
+    return Z.valueOf(n * (long) (n + 1)).multiply((3L * k + 2) * n + 1).divide(6);
   }
 }
