+++ src/irvine/oeis/a155/A155955.java
@@ -11,6 +11,6 @@ public class A155955 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return Z.valueOf(n * k).pow(k);
+    return Z.valueOf(n * (long) k).pow(k);
   }
 }
