+++ src/irvine/oeis/a269/A269945.java
@@ -12,6 +12,6 @@ public class A269945 extends Triangle {
 
   @Override
   public Z compute(final int n, final int k) {
-    return n == 0 ? Z.ONE : get(n - 1, k - 1).add(get(n - 1, k).multiply(k * k));
+    return n == 0 ? Z.ONE : get(n - 1, k - 1).add(get(n - 1, k).multiply(k * (long) k));
   }
 }
